add Config MENGINE_INLINE add MENGINE_ASSERTION_PYTHON_EVENT_RECEIVER remove BaseUpdation _update wip plugin Spine fix plugin Texturepacker

+++ src/Plugins/TexturepackerPlugin/ResourceTexturepacker.cpp
@@ -26,12 +26,12 @@ namespace Mengine
     {
     }
     //////////////////////////////////////////////////////////////////////////
-    float ResourceTexturepacker::getAtlasWidth() const
+    uint32_t ResourceTexturepacker::getAtlasWidth() const
     {
         return m_atlasWidth;
     }
     //////////////////////////////////////////////////////////////////////////
-    float ResourceTexturepacker::getAtlasHeight() const
+    uint32_t ResourceTexturepacker::getAtlasHeight() const
     {
         return m_atlasHeight;
     }
